+++ src/components/Attachment/UnsupportedAttachment.tsx
@@ -1,4 +1,6 @@
 import React from 'react';
+import { FileIcon } from '../ReactFileUtilities';
+import { useTranslationContext } from '../../context';
 import type { Attachment } from 'stream-chat';
 import type { DefaultStreamChatGenerics } from '../../types/types';
 
@@ -12,11 +14,19 @@ export const UnsupportedAttachment = <
   StreamChatGenerics extends DefaultStreamChatGenerics = DefaultStreamChatGenerics
 >({
   attachment,
-}: UnsupportedAttachmentProps<StreamChatGenerics>) => (
-  <div>
-    <div>
-      Unsupported attachment type <strong>{attachment.type ?? 'unknown'}</strong>
+}: UnsupportedAttachmentProps<StreamChatGenerics>) => {
+  const { t } = useTranslationContext('UnsupportedAttachment');
+  return (
+    <div className='str-chat__message-attachment-unsupported' data-testid='attachment-unsupported'>
+      <FileIcon className='str-chat__file-icon' />
+      <div className='str-chat__message-attachment-unsupported__metadata'>
+        <div
+          className='str-chat__message-attachment-unsupported__title'
+          data-testid='unsupported-attachment-title'
+        >
+          {attachment.title || t<string>('Unsupported attachment')}
+        </div>
+      </div>
     </div>
-    <code>{JSON.stringify(attachment, null, 4)}</code>;
-  </div>
-);
+  );
+};