+++ src/components/MessageInput/hooks/useSubmitHandler.ts
@@ -102,9 +102,9 @@ export const useSubmitHandler = <
       }));
 
     const otherAttachments = attachments
-      .filter((att) => att.$internal?.uploadState !== 'failed')
+      .filter((att) => att.localMetadata?.uploadState !== 'failed')
       .map((localAttachment) => {
-        const { $internal: _, ...attachment } = localAttachment;
+        const { localMetadata: _, ...attachment } = localAttachment;
         return attachment as Attachment;
       });
 
@@ -133,7 +133,7 @@ export const useSubmitHandler = <
     const someAttachmentsUploading =
       Object.values(imageUploads).some((upload) => upload.state === 'uploading') ||
       Object.values(fileUploads).some((upload) => upload.state === 'uploading') ||
-      attachments.some((att) => att.$internal?.uploadState === 'uploading');
+      attachments.some((att) => att.localMetadata?.uploadState === 'uploading');
 
     if (someAttachmentsUploading) {
       return addNotification(t('Wait until all attachments have uploaded'), 'error');